requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
slowapi
email-validator
python-dotenv
